Split out rrr_game.rs into separate files
