+++ rust-server/src/rrr_game/mod.rs
@@ -1,0 +1,14 @@
+// GAME_NAME:
+const GAME_NAME: &str = "rrr-game";
+const CHUNK_LENGTH: usize = 9;
+
+mod create;
+pub use create::create_game;
+
+mod coord;
+
+mod get;
+pub use get::get_gamestate;
+
+mod action;
+pub use action::do_action;
